rename schedule maker to occupancy
schedule maker is making occupancies. Hence, it is renamed as occupancy helper
