+++ cea/default.config
@@ -44,12 +44,8 @@ databases.help = List of databases to initialize from the selected path
 input-databases = comfort, architecture, air-conditioning, internal-loads, supply, schedules
 input-databases.type = MultiChoiceParameter
 input-databases.choices = comfort, architecture, air-conditioning, internal-loads, supply, schedules
-input-databases.help = List of inputs to map from your databases to your new/existing scenario
-
-buildings =
-buildings.type = BuildingsParameter
-buildings.help = List of buildings to create schedules for (leave blank to consider all).
-buildings.category = Specific inputs for the schedule database
+input-databases.help = List of inputs to map from Database to each building in Scenario.
+input-databases.category = Advanced
 
 [streets-helper]
 bbox =
@@ -268,15 +264,16 @@ buffer = 50
 buffer.type = RealParameter
 buffer.help = Perimeter buffer (m) around sample buildings .
 
-[schedule-maker]
-buildings =
-buildings.type = BuildingsParameter
-buildings.help = List of buildings considered for the schedule model (to simulate all buildings leave blank).
+[occupancy-helper]
 
 schedule-model = deterministic
 schedule-model.type = ChoiceParameter
 schedule-model.choices = deterministic, stochastic
-schedule-model.help = Type of schedule model to use (stochastic or deterministic)
+schedule-model.help = Type of occupancy model to use.
+
+buildings =
+buildings.type = BuildingsParameter
+buildings.help = Lis of buildings for building occupancy generation. Leave blank to select all.
 
 [demand]
 buildings =
