+++ cea/tests/test_calc_thermal_loads.py
@@ -8,7 +8,7 @@ import unittest
 
 import pandas as pd
 
-from cea.demand.schedule_maker.schedule_maker import schedule_maker_main
+from cea.demand.occupancy_helper import schedule_maker_main
 from cea.demand.building_properties import BuildingProperties
 from cea.demand.thermal_loads import calc_thermal_loads
 from cea.utilities.date import get_date_range_hours_from_year
